// src/enums/menu.enum.ts
export enum MainMenuCategories {
  MainMenu = 'Main Menu',
  HotCoffees = 'Hot Coffees',
  HotDrinks = 'Hot Drinks',
  ColdCoffees = 'Cold Coffees',
  ColdDrinks = 'Cold Drinks',
  HotTeas = 'Hot Teas',
  IcedTeas = 'Iced Teas',
  HotBreakfest = 'Hot Breakfast',
  Lunch = 'Lunch',
  YogurtCustard = 'Yogurt & Custard',
  Bakery = 'Bakery',
  SnacksAndSweets = 'Snacks and Sweets',
  WholeBean = 'Whole Bean',
  VIAInstant = 'VIA Instant',
  VerismoPods = 'Verismo Pods',
  ColdBrew = 'Cold Brew'
}